requirements.txt
aiofiles==23.1.0
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
anyio==3.6.2
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
autopep8==2.0.2
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
branca==0.6.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cryptography==41.0.4
cssselect2==0.7.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.6
dj-database-url==2.0.0
Django==4.2.7
django-allauth==0.57.0
django-bootstrap-v5==1.0.11
django-bootstrap4==23.2
django-crispy-forms==2.1
django-heroku==0.3.1
django-map-widgets==0.4.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.95.1
filelock==3.9.0
folium==0.14.0
fonttools==4.39.3
fpdf==1.7.2
frozenlist==1.4.0
gogoanimepy==0.0.3
google-api-core==2.11.1
google-api-python-client==2.90.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
gunicorn==20.1.0
h11==0.14.0
html5lib==1.1
httplib2==0.22.0
idna==3.4
img2pdf==0.4.4
import-export==0.3.1
ipykernel==6.23.0
ipython==8.13.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
multidict==6.0.4
mysqlclient==2.2.0
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.24.2
oauthlib==3.2.2
opencage==2.3.0
openpyxl==3.1.2
oscrypto==1.3.0
outcome==1.2.0
packaging==23.1
pandas==2.0.1
parso==0.8.3
pbr==5.11.1
phonenumbers==8.13.19
pickleshare==0.7.5
pikepdf==7.2.0
Pillow==9.5.0
platformdirs==3.1.1
plotly==5.14.1
pornhub-api==0.3.0
pornhubapi==0.1.0
postgis==1.0.4
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
psycopg2==2.9.6
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.1
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
PyJWT==2.8.0
pyparsing==3.0.9
pypdf==3.16.2
pypng==0.20220715.0
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
python-multipart==0.0.6
python3-openid==3.2.0
pytz==2023.3
pywin32==306
PyYAML==6.0.1
pyzmq==25.0.2
qrcode==7.4.2
regex==2023.3.23
reportlab==3.6.13
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sc-google-maps-api==1.0.4
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
selenium==4.12.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.3
stack-data==0.6.2
starlette==0.26.1
stevedore==5.0.0
svglib==1.5.1
tenacity==8.2.2
threadpoolctl==3.1.0
tinycss2==1.2.1
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.10.4
twilio==8.9.1
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==5.1
uritemplate==4.1.1
uritools==4.0.2
urllib3==1.26.16
uvicorn==0.22.0
virtualenv==20.20.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wcwidth==0.2.6
webencodings==0.5.1
whitenoise==6.4.0
wsproto==1.2.0
xhtml2pdf==0.2.11
yarl==1.9.2
yellowbrick==1.5
